requirements.txt
click>=8.0
GitPython>=3.1
PyYAML>=6.0
Jinja2>=3.0
colorama>=0.4
tabulate>=0.9